requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
av==14.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
ctranslate2==4.5.0
debugpy==1.8.13
deepl==1.16.1 # Add deepl library
decorator==5.2.1
distro==1.9.0
executing==2.2.0
faster-whisper==1.1.1
ffmpeg-python==0.2.0
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.1
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.6.0
mpmath==1.3.0
narwhals==1.32.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
onnxruntime==1.21.0
openai==1.70.0
openai-whisper==20240930
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic_core==2.33.0
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.23.1
rsa==4.9
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
srt==3.5.3
stack-data==0.6.3
streamlit==1.43.2
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
yt-dlp==2025.2.19
